+++ src/main/java/com/create/chacha/domains/buyer/exception/classes/ReservationException.java
@@ -1,6 +1,4 @@
-package com.create.chacha.domains.buyer.exception;
-
-import com.create.chacha.common.constants.ResponseCode;
+package com.create.chacha.domains.buyer.exception.classes;
 
 public class ReservationException extends RuntimeException{
     public ReservationException(String message) {
