+++ src/main/java/com/create/chacha/domains/buyer/exception/classes/ReservationSaveException.java
@@ -1,4 +1,4 @@
-package com.create.chacha.domains.buyer.exception;
+package com.create.chacha.domains.buyer.exception.classes;
 
 public class ReservationSaveException extends RuntimeException{
     public ReservationSaveException(String message, Throwable cause) {
