+++ src/main/java/com/create/chacha/domains/buyer/exception/payment/PaymentFailedException.java
@@ -1,4 +1,4 @@
-package com.create.chacha.domains.buyer.exception;
+package com.create.chacha.domains.buyer.exception.payment;
 
 public class PaymentFailedException extends RuntimeException {
     public PaymentFailedException(String message) {
