+++ src/main/java/com/create/chacha/domains/buyer/exception/payment/PaymentRequestException.java
@@ -1,4 +1,4 @@
-package com.create.chacha.domains.buyer.exception;
+package com.create.chacha.domains.buyer.exception.payment;
 
 public class PaymentRequestException extends RuntimeException {
     public PaymentRequestException(String message, Throwable cause) {
